backend/requirements.txt
Flask>=2.0
flask-cors
boto3
botocore
opensearch-py
pypdf
requests-aws4auth
SQLAlchemy>=1.4
psycopg2-binary  # if using PostgreSQL
python-dotenv
